Refactor: I improved the n-tier architecture by moving the JWT logic to a dedicated Session layer. I implemented the Session layer to handle JWT token generation, removed this responsibility from the AuthController in the Services layer, and refactored the login endpoint to return a JWT token, aligning the design with a stricter and cleaner n-tier architecture.
