+++ src/Session/ITokenService.cs
@@ -1,0 +1,7 @@
+namespace Session
+{
+    public interface ITokenService
+    {
+        string GenerateJwtToken(string username);
+    }
+}
